+++ Services/WebApp/src/routes/(session)/(sidebar)/admin/users/+page.server.ts
@@ -1,48 +1,60 @@
 import { UserDatabase, type User } from '$lib/database/userDatabase';
+import { redirect } from '@sveltejs/kit';
 import { ObjectId } from 'mongodb';
 import type { Actions, PageServerLoad } from './$types';
-import { redirect } from '@sveltejs/kit';
 
 export const load = (async () => {
-    let users: User[] = await (await UserDatabase.get()).getAllUsers();
+	let users: User[] = await (await UserDatabase.get()).getAllUsers();
 
-    return {
-        users: structuredClone(users.map(user => ({ ...user, _id: user._id?.toString(), userId: user.userId?.toString() })))
-    };
+	return {
+		users: structuredClone(
+			users.map((user) => ({ ...user, _id: user._id?.toString(), userId: user.userId?.toString() }))
+		)
+	};
 }) satisfies PageServerLoad;
 
 export const actions = {
-    remove_role: async ({ request }) => {
-        const data: FormData = await request.formData();
-        const role: string = data.get('role') as string;
-        const user_id: string = data.get('user_id') as string;
+	remove_role: async ({ request, url }) => {
+		const data: FormData = await request.formData();
+		const role: string = data.get('role') as string;
+		const user_id: string = data.get('user_id') as string;
 
-        await UserDatabase.get().then((userDatabase) => {
-            userDatabase.deleteRole(new ObjectId(user_id), role);
-        });
+		await UserDatabase.get().then((userDatabase) => {
+			userDatabase.deleteRole(new ObjectId(user_id), role);
+		});
 
-        throw redirect(302, '/admin/users');
-    },
-    add_role: async ({ request }) => {
-        const data: FormData = await request.formData();
-        const role: string = data.get('role') as string;
-        const user_id: string = data.get('user_id') as string;
+		throw redirect(302, url.searchParams.get('redirect_url') || '/admin/users');
+	},
+	add_role: async ({ request, url }) => {
+		const data: FormData = await request.formData();
+		const role: string = data.get('role') as string;
+		const user_id: string = data.get('user_id') as string;
 
-        await UserDatabase.get().then((userDatabase) => {
-            userDatabase.addRole(new ObjectId(user_id), role);
-        });
+		await UserDatabase.get().then((userDatabase) => {
+			userDatabase.addRole(new ObjectId(user_id), role);
+		});
 
-        throw redirect(302, '/admin/users');
-    },
-    add_user: async ({ request }) => {
-        const data: FormData = await request.formData();
-        const email: string = data.get('email') as string;
-        const roles: string[] = data.getAll('roles') as string[];
+		throw redirect(302, url.searchParams.get('redirect_url') || '/admin/users');
+	},
+	add_user: async ({ request, url }) => {
+		const data: FormData = await request.formData();
+		const email: string = data.get('email') as string;
+		const roles: string[] = data.getAll('roles') as string[];
 
-        await UserDatabase.get().then((userDatabase) => {
-            userDatabase.addUser({ email, roles });
-        });
+		await UserDatabase.get().then((userDatabase) => {
+			userDatabase.addUser({ email, roles });
+		});
 
-        throw redirect(302, '/admin/users');
-    }
+		throw redirect(302, url.searchParams.get('redirect_url') || '/admin/users');
+	},
+	delete_user: async ({ request, url }) => {
+		const data: FormData = await request.formData();
+		const user_id: string = data.get('user_id') as string;
+
+		await UserDatabase.get().then((userDatabase) => {
+			userDatabase.deleteUser(new ObjectId(user_id));
+		});
+
+		throw redirect(302, url.searchParams.get('redirect_url') || '/admin/users');
+	}
 } satisfies Actions;
